add Account test + css 90%
