+++ src/main/java/com/example/wk/AccountDTO.java
@@ -1,0 +1,15 @@
+package com.example.wk;
+
+import lombok.Data;
+
+@Data
+public class AccountDTO {
+	private String a_name;
+	private String a_id;
+	private String a_pw;
+	private String a_gender;
+	private String a_addr;
+	private String a_interest;
+	private String a_introduce;
+	private String a_img;
+}
