+++ src/main/java/com/example/wk/Human.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.wk;
 
 import lombok.Data;
