+++ src/main/java/com/example/wk/HumanC.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.wk;
 
 import java.util.List;
 
@@ -45,17 +45,15 @@ public class HumanC {
 		return "NewFile";
 	}
 	
-	/* */
-	
-	@GetMapping("artist_reg")
-	public String artist_reg() {
-		return "artist_reg";
-	}
-	
-	@GetMapping("artist_update")
-	public String artist_update() {
-		return "artist_update";
-	}
+//	@GetMapping("artist_reg")
+//	public String artist_reg() {
+//		return "artist_reg";
+//	}
+//	
+//	@GetMapping("artist_update")
+//	public String artist_update() {
+//		return "artist_update";
+//	}
 	
 //	@GetMapping("artist_detail_viewer")
 //	public String artist_detail_viewer(Model model) {
@@ -63,31 +61,28 @@ public class HumanC {
 //		return "artist_detail_viewer";
 //	}
 	
-	@GetMapping("artist_detail_viewer")
-	public String artist_detail_viewer() {
-	    return "artist_detail_viewer";
-	}
-	
-	@GetMapping("artist_detail_owner")
-	public String artist_detail_owner() {
-		return "artist_detail_owner";
-	}
-	
-	@GetMapping("/music_upload")
-	public String music_upload() {
-		return "music_upload";
-	}
-	
-	@GetMapping("/mypage_artist")
-	public String mypage_artist() {
-		return "mypage_artist";
-	}
-	
-	@GetMapping("/mypage_user")
-	public String mypage_user() {
-		return "mypage_user";
-	}
-	
-	
-	
+//	@GetMapping("artist_detail_viewer")
+//	public String artist_detail_viewer() {
+//	    return "artist_detail_viewer";
+//	}
+//	
+//	@GetMapping("artist_detail_owner")
+//	public String artist_detail_owner() {
+//		return "artist_detail_owner";
+//	}
+//	
+//	@GetMapping("/music_upload")
+//	public String music_upload() {
+//		return "music_upload";
+//	}
+//	
+//	@GetMapping("/mypage_artist")
+//	public String mypage_artist() {
+//		return "mypage_artist";
+//	}
+//	
+//	@GetMapping("/mypage_user")
+//	public String mypage_user() {
+//		return "mypage_user";
+//	}
 }
