+++ src/main/java/com/example/wk/HumanMapper.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.wk;
 
 import java.util.List;
