+++ src/main/java/com/example/wk/SemiApplication.java
@@ -1,11 +1,13 @@
-package com.example.demo;
+package com.example.wk;
+
+
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan(basePackages = "com.example.demo")
+@ComponentScan(basePackages = "com.example.wk")
 public class SemiApplication {
 
 	public static void main(String[] args) {
