+++ src/main/java/com/example/wk/TestService_Model.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.wk;
 
 import java.util.List;
