+++ src/main/resources/application.properties
@@ -13,4 +13,4 @@ spring.datasource.password=Soldesk802!!
 mybatis.mapper-locations=classpath:mappers/*.xml
 
 # xml파일 result type에 패키지명을 생략할 수 있도록 alias 설정
-mybatis.type-aliases-package=com.example.demo
+mybatis.type-aliases-package=com.example.wk
