+++ src/main/resources/mappers/account.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.example.demo.AccountMapper">
+  <select id="getAccount" resultType="AccountDTO">
+    select * from account_test where a_name = #{val}
+  </select>
+  <select id="getAllAccount" resultType="AccountDTO" >
+    select * from account_test
+  </select>
+  <insert id="regAccount">
+	  insert into account_test values(#{a_name}, #{a_age})
+  </insert>
+</mapper>
