+++ src/main/resources/static/css/music_upload.css
@@ -60,6 +60,10 @@ button {
     margin: 10px;
 }
 
+button:hover {
+	background-color: #555;
+}
+
 .button-box{
 	width: 100%;
 	display: flex;
